fix(room): address CodeRabbit review comments
- Move -last_activity before status in order_by and add pk tie-breaker
  so newest rooms always appear first regardless of status group
- Split test_room_list_ordering.py into two files (one class per file)
  to satisfy the static-analysis CI check
- Remove unused db fixture arg from test method (ruff ARG002)
