+++ apps/room/tests/test_services/test_room_qs_for_list_ordering.py
@@ -5,27 +5,8 @@ from django.utils import timezone
 
 from apps.room.models import Room
 from apps.room.tests.factories import RoomFactory
+from apps.transaction.models import ParentTransaction
 from apps.transaction.tests.factories import ParentTransactionFactory
-
-
-@pytest.mark.django_db
-class TestAnnotateLastActivity:
-    def test_room_with_transaction_uses_transaction_timestamp(self, room, user):
-        transaction = ParentTransactionFactory(room=room, paid_by=user, currency=room.preferred_currency)
-        expected_ts = timezone.now() - timedelta(days=1)
-        from apps.transaction.models import ParentTransaction
-
-        ParentTransaction.objects.filter(pk=transaction.pk).update(lastmodified_at=expected_ts)
-
-        annotated = Room.objects.filter(pk=room.pk).annotate_last_activity().get()
-        # Should be close to expected_ts (within a second for DB rounding)
-        assert abs((annotated.last_activity - expected_ts).total_seconds()) < 2
-
-    def test_room_without_transactions_falls_back_to_room_lastmodified_at(self, db):
-        room = RoomFactory()
-        annotated = Room.objects.filter(pk=room.pk).annotate_last_activity().get()
-        assert annotated.last_activity is not None
-        assert abs((annotated.last_activity - room.lastmodified_at).total_seconds()) < 2
 
 
 @pytest.mark.django_db
@@ -38,17 +19,12 @@ class TestRoomQsForListOrdering:
         new_room = RoomFactory(created_by=user)
         new_room.users.add(user)
 
-        # Give old_room a transaction with an old timestamp
         old_tx = ParentTransactionFactory(room=old_room, paid_by=user, currency=old_room.preferred_currency)
-        from apps.transaction.models import ParentTransaction
-
         ParentTransaction.objects.filter(pk=old_tx.pk).update(lastmodified_at=now - timedelta(days=10))
 
-        # Give new_room a transaction with a recent timestamp
         new_tx = ParentTransactionFactory(room=new_room, paid_by=user, currency=new_room.preferred_currency)
         ParentTransaction.objects.filter(pk=new_tx.pk).update(lastmodified_at=now - timedelta(days=1))
 
-        # Bust the cached_property so we get a fresh queryset
         if "room_qs_for_list" in user.__dict__:
             del user.__dict__["room_qs_for_list"]
 
@@ -71,12 +47,7 @@ class TestRoomQsForListOrdering:
         tx_room = RoomFactory(created_by=user)
         tx_room.users.add(user)
         tx = ParentTransactionFactory(room=tx_room, paid_by=user, currency=tx_room.preferred_currency)
-        from apps.transaction.models import ParentTransaction
-
         ParentTransaction.objects.filter(pk=tx.pk).update(lastmodified_at=now - timedelta(days=1))
-
-        # Force no_tx_room to be older so its lastmodified_at fallback sorts it after tx_room
-        from apps.room.models import Room
 
         Room.objects.filter(pk=no_tx_room.pk).update(lastmodified_at=now - timedelta(days=30))
